Attempted to address "unused function parameter" errors by adding skeleton implementations for deinit functions
Implemented placeholder versions of deinit functions for each AstNode variant. These stubs currently do not perform any actual memory deallocation but were added to satisfy the compiler's requirement for function parameters to be used. This step was taken in an attempt to resolve the compilation errors related to unused function parameters.

⛔️ Error: Actual memory management logic is yet to be implemented in these deinit functions.